backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
Werkzeug==2.3.7
PyJWT==2.9.0